fix(challenges): change test for template literals challenge
The original test was too rigid, and many users were forced to use one solution, usually map, to
complete the challenge. The test is changed here to allow for multiple solutions to the problem,
including the use of either single or double quotes when naming a class.

BREAKING CHANGE:
None