fix(schema): change schema and unpack script to fix tests in main repo
Added an additional argument to the getChallenges function call in unpack.js to act as a flag for
when the unpack script is running. Then modified the lodash omit array so that translations are
included in the array when the test suite is running, but excluded from the array when the unpack
script is running. Effectively this means that the translations are not included in the challenge
block when going through the test suite, but included in the challenge block for the unpack script.
Finally I removed the unnecessary releasedOn section in unpackedChallenge.js, and removed both
releasedOn and translations from challengeSchema.js.

All of these changes should fix the tests
over on the main fCC branch.

ISSUES CLOSED: Partially addresses #204